Refactor

+++ point.go
@@ -1,10 +1,10 @@
 package fractal
 
 type Point struct {
-	C complex128
-	X int
-	Y int
-	Escape float64
-	FinalZ complex128
+	C       complex128
+	X       int
+	Y       int
+	Escape  int
+	FinalZ  complex128
 	Escaped bool
- }
+}
